requirements.txt
appnope==0.1.0
arrow==0.15.5
asgiref==3.2.7
backcall==0.1.0
binaryornot==0.4.4
blessings==1.7
boto3==1.11.1
botocore==1.14.17
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cookiecutter==1.7.0
coverage==5.0.3
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.3
django-configurations==2.2
django-filter==2.2.0
django-model-utils==4.0.0
django-nose==1.4.6
django-storages==1.8
django-unique-upload==0.2.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.3
flake8==3.7.9
future==0.18.2
gunicorn==20.0.4
idna==2.9
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.9.5
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.0.4
mock==3.0.5
newrelic==5.4.1.134
nose==1.3.7
nose-progressive==1.5.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
whichcraft==0.6.1
